+++ packages/connector-github/src/index.ts
@@ -1,0 +1,4 @@
+export * from './client.js';
+export * from './errors.js';
+export * from './tools.js';
+export * from './environment.js';